# api-docs/api-docs.yaml
openapi: 3.0.3
info:
  title: PizzaFast API REST
  version: v1
  description: |
    API REST de PizzaFast para autenticacion, usuarios, catalogo, imagenes y pedidos.
    La autenticacion se basa en cookies HTTP-only generadas por los endpoints de auth.
servers:
  - url: https://localhost:8443
    description: Desarrollo local
...
paths:
...
  /api/v1/users/register:
    post:
      tags: [Users]
      summary: Registrar usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterDTO'
      responses:
        '201':
          description: Usuario registrado
        '400':
          description: Datos invalidos
  /api/v1/users/me:
    get:
      tags: [Users]
      summary: Obtener mi perfil
      security:
        - accessTokenCookie: []
      responses:
        '200':
          description: Perfil obtenido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '401':
          description: No autenticado
    put:
      tags: [Users]
      summary: Actualizar mi perfil
      security:
        - accessTokenCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateDTO'
      responses:
        '200':
          description: Perfil actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: Datos invalidos
        '401':
          description: No autenticado
  /api/v1/users/me/orders:
    get:
      tags: [Users]
      summary: Obtener mis pedidos
      security:
        - accessTokenCookie: []
      responses:
        '200':
          description: Pedidos obtenidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderDTO'
        '401':
          description: No autenticado
  /api/v1/users/:
    get:
      tags: [Users]
      summary: Listar usuarios
      security:
        - accessTokenCookie: []
      responses:
        '200':
          description: Usuarios obtenidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTO'
        '403':
          description: Acceso denegado
  /api/v1/users/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Users]
      summary: Obtener usuario por ID
      security:
        - accessTokenCookie: []
      responses:
        '200':
          description: Usuario encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '404':
          description: Usuario no encontrado
        '403':
          description: Acceso denegado
    delete:
      tags: [Users]
      summary: Eliminar usuario
      security:
        - accessTokenCookie: []
      responses:
        '204':
          description: Usuario eliminado
        '404':
          description: Usuario no encontrado
        '403':
          description: Acceso denegado
...
components:
  securitySchemes:
    accessTokenCookie:
      type: apiKey
      in: cookie
      name: AccessToken
      description: JWT de acceso almacenado en cookie HTTP-only.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
...
    UserRegisterDTO:
      type: object
      required: [name, email, password, confirmPassword]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        confirmPassword:
          type: string
    UserUpdateDTO:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        oldPassword:
          type: string
        newPassword:
          type: string
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        name:
          type: string
        roles:
          type: array
          items:
            type: string
...
    OrderDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        address:
          type: string
        city:
          type: string
        postalCode:
          type: string
        phoneNumber:
          type: string
        userEmail:
          type: string
          format: email
        productTitles:
          type: array
          items:
            type: string
...